Split reducers into own files.
